+++ src/dplyr.cpp
@@ -1817,8 +1817,7 @@ SEXP mutate_impl( DataFrame df, List args, Environment env){
         return mutate_grouped<RowwiseDataFrame, LazyRowwiseSubsets>( df, args, dots);
     } else if( is<GroupedDataFrame>( df ) ){
         // handle special case where there are no groups
-        IntegerVector group_sizes = df.attr("group_sizes" );
-        if( !group_sizes.size() ){
+        if( df.nrows() == 0 ){
             DataFrame res = mutate_not_grouped( df, args, dots) ;
             res.attr("vars") = df.attr("vars" ) ;
             return GroupedDataFrame(res).data() ;